REF: Created pandas.core.arrays (#19269)
Moved pandas.core.categorical to arrays.
