+++ pandas/core/arrays/__init__.py
@@ -1,0 +1,1 @@
+from .categorical import Categorical  # noqa
