+++ pandas/core/sorting.py
@@ -182,7 +182,7 @@ def indexer_from_factorized(labels, shape, compress=True):
 
 
 def lexsort_indexer(keys, orders=None, na_position='last'):
-    from pandas.core.categorical import Categorical
+    from pandas.core.arrays import Categorical
 
     labels = []
     shape = []
