+++ pandas/tests/series/test_api.py
@@ -511,8 +511,7 @@ class TestCategoricalSeries(object):
 
     def test_cat_accessor_api(self):
         # GH 9322
-        from pandas.core.categorical import CategoricalAccessor
-
+        from pandas.core.arrays.categorical import CategoricalAccessor
         assert Series.cat is CategoricalAccessor
         s = Series(list('aabbcde')).astype('category')
         assert isinstance(s.cat, CategoricalAccessor)
